+++ ReleaseNotes.md
@@ -1,4 +1,8 @@
 # 3.0.4xx (Caba series release 8)
+* [BUGFIX] for `-c|callContext` option -- generate valid IL where the context function completes with a tail call
+* [BUGFIX] for `-c|callContext` option -- generate valid IL where the context function contains a branch directly to a return instruction
+* `-c|cobertura` option in `runner` mode (parameter `Cobertura` for the `AltCover.Collect` task) -- Cobertura format output to the given file name, with format composition based on the sample file at https://raw.githubusercontent.com/jenkinsci/cobertura-plugin/master/src/test/resources/hudson/plugins/cobertura/coverage-with-data.xml
+
 
 # 3.0.433 (Caba series release 7)
 * `-t|threshold` option in `runner` mode (parameter `Threshold` for the `AltCover.Collect` task) to fail the build (non-zero return code or MSBuild error state) if coverage falls below the specified percentage